Return Float instead of Double

+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/ui/gesture/Drag.kt
@@ -44,7 +44,7 @@ interface Drag {
  * - 12 o'clock = 0 degrees
  * - 3 o'clock = 90 degrees
  */
-fun dragAngleDegrees(delta: Offset): Double  =
+fun dragAngleDegrees(delta: Offset): Float  =
     angleDegrees(delta)
 
 /**
